Implementing N-Tier Data Layer with Protos and Models in shared library, consumed by gRPC API Service.
