+++ hello-dotnet6/HelloDotNet6/HelloDotNet6.Data/Class1.cs
@@ -1,0 +1,7 @@
+﻿namespace HelloDotNet6.Data
+{
+    public class Class1
+    {
+
+    }
+}
