+++ hello-dotnet6/HelloDotNet6/HelloDotNet6.GrpcApi/Properties/launchSettings.json
@@ -1,0 +1,18 @@
+{
+  "profiles": {
+    "HelloDotNet6.GrpcApi": {
+      "commandName": "Project",
+      "environmentVariables": {
+        "ASPNETCORE_ENVIRONMENT": "Development"
+      },
+      "applicationUrl": "http://localhost:5046;https://localhost:7046",
+      "dotnetRunMessages": true
+    },
+    "Docker": {
+      "commandName": "Docker",
+      "launchUrl": "{Scheme}://{ServiceHost}:{ServicePort}",
+      "publishAllPorts": true,
+      "useSSL": true
+    }
+  }
+}
