+++ hello-dotnet6/HelloDotNet6/HelloDotNet6.sln
@@ -5,11 +5,13 @@ VisualStudioVersion = 17.2.32414.248
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Console", "Console\Console.csproj", "{DB5AE002-0D45-4D5A-8C76-81813B2EE11D}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ServiceAgRPC", "ServiceAgRPC\ServiceAgRPC.csproj", "{A9BA2EE2-E614-4C70-88AF-261351935A30}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebAPIAhttp", "WebAPIAhttp\WebAPIAhttp.csproj", "{A1CC4072-29FF-40E6-8E7B-B23E53FFA7E5}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SharedLib", "SharedLib\SharedLib.csproj", "{78A626A3-E6DF-4873-8EA4-07337DDDC305}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SharedLib", "SharedLib\SharedLib.csproj", "{78A626A3-E6DF-4873-8EA4-07337DDDC305}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HelloDotNet6.GrpcApi", "HelloDotNet6.GrpcApi\HelloDotNet6.GrpcApi.csproj", "{DD81F41D-DCE4-41FC-A1F3-E2C57B70B374}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HelloDotNet6.Data", "HelloDotNet6.Data\HelloDotNet6.Data.csproj", "{6D75FD5C-70C5-4675-B703-B62235E27BE2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,10 +23,6 @@ Global
 		{DB5AE002-0D45-4D5A-8C76-81813B2EE11D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DB5AE002-0D45-4D5A-8C76-81813B2EE11D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DB5AE002-0D45-4D5A-8C76-81813B2EE11D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A9BA2EE2-E614-4C70-88AF-261351935A30}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A9BA2EE2-E614-4C70-88AF-261351935A30}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A9BA2EE2-E614-4C70-88AF-261351935A30}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A9BA2EE2-E614-4C70-88AF-261351935A30}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A1CC4072-29FF-40E6-8E7B-B23E53FFA7E5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A1CC4072-29FF-40E6-8E7B-B23E53FFA7E5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A1CC4072-29FF-40E6-8E7B-B23E53FFA7E5}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -33,6 +31,14 @@ Global
 		{78A626A3-E6DF-4873-8EA4-07337DDDC305}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{78A626A3-E6DF-4873-8EA4-07337DDDC305}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{78A626A3-E6DF-4873-8EA4-07337DDDC305}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DD81F41D-DCE4-41FC-A1F3-E2C57B70B374}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DD81F41D-DCE4-41FC-A1F3-E2C57B70B374}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DD81F41D-DCE4-41FC-A1F3-E2C57B70B374}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DD81F41D-DCE4-41FC-A1F3-E2C57B70B374}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6D75FD5C-70C5-4675-B703-B62235E27BE2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6D75FD5C-70C5-4675-B703-B62235E27BE2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6D75FD5C-70C5-4675-B703-B62235E27BE2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6D75FD5C-70C5-4675-B703-B62235E27BE2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
