--- hello-dotnet6/HelloDotNet6/ServiceAgRPC/Properties/launchSettings.json
@@ -1,18 +1,0 @@
-{
-  "profiles": {
-    "ServiceAgRPC": {
-      "commandName": "Project",
-      "environmentVariables": {
-        "ASPNETCORE_ENVIRONMENT": "Development"
-      },
-      "applicationUrl": "http://localhost:5096;https://localhost:7096",
-      "dotnetRunMessages": true
-    },
-    "Docker": {
-      "commandName": "Docker",
-      "launchUrl": "{Scheme}://{ServiceHost}:{ServicePort}",
-      "publishAllPorts": true,
-      "useSSL": true
-    }
-  }
-}
