--- hello-dotnet6/HelloDotNet6/ServiceAgRPC/Protos/greet.proto
@@ -1,21 +1,0 @@
-syntax = "proto3";
-
-option csharp_namespace = "ServiceAgRPC";
-
-package greet;
-
-// The greeting service definition.
-service Greeter {
-  // Sends a greeting
-  rpc SayHello (HelloRequest) returns (HelloReply);
-}
-
-// The request message containing the user's name.
-message HelloRequest {
-  string name = 1;
-}
-
-// The response message containing the greetings.
-message HelloReply {
-  string message = 1;
-}
