--- hello-dotnet6/HelloDotNet6/ServiceAgRPC/Services/GreeterService.cs
@@ -1,22 +1,0 @@
-using Grpc.Core;
-using ServiceAgRPC;
-
-namespace ServiceAgRPC.Services
-{
-    public class GreeterService : Greeter.GreeterBase
-    {
-        private readonly ILogger<GreeterService> _logger;
-        public GreeterService(ILogger<GreeterService> logger)
-        {
-            _logger = logger;
-        }
-
-        public override Task<HelloReply> SayHello(HelloRequest request, ServerCallContext context)
-        {
-            return Task.FromResult(new HelloReply
-            {
-                Message = "Hello " + request.Name
-            });
-        }
-    }
-}
